Hex Scan: Pokepath Mod (#1010)
The "Pokepath" changes the order of the scan points for the hex method so that no scan ever "teleports" more than one scan point away. Check the PR for a visual diagram of what this looks like (and why it's called the "pokepath").

The code will also not start at the center location, but a few steps rolled back so that the central 2 or 3 steps are fully covered before proceeding to the northern segment of the scan. This way you get a relatively decent coverage of the "central scan nugget" and the full path benefit.