Update route path and popups
Connect first waypoint to route start and last waypoint to route end.  This will look different than in the game, but should make more sense on the map.

Set popup info to be the same between start, end, and route path.

Fixed path popup location to match mouse pointer, also allow it to move to follow the mouse.